ASTERIX: Widen a loop variable
Fix

    /wireshark/epan/dissectors/packet-asterix.c:61495:35: error: loop variable has narrower type 'uint8_t' than iteration's upper bound 'const unsigned int' [bugprone-too-small-loop-variable,-warnings-as-errors]
     61495 |             for (i = 0, size = 0; i < field->header_length; i++) {
           |                                   ^

(cherry picked from commit 6a6f5da68ef77ea285acaf6d75142b6c9aa5d62f)

+++ epan/dissectors/packet-asterix.c
@@ -59178,7 +59178,7 @@ static int asterix_field_length (tvbuff_t *tvb, unsigned offset, const AsterixFi
     unsigned bit_size;
     unsigned size;
     uint64_t count;
-    uint8_t i;
+    unsigned i;
     bool should_break;
 
     size = 0;
